moved server into its own file
